Fixes norminette errors

+++ utils/ft_printf_utils.c
@@ -6,7 +6,7 @@
 /*   By: icezar-s <icezar-s@student.42porto.co      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/25 16:11:53 by icezar-s          #+#    #+#             */
-/*   Updated: 2025/12/06 17:02:24 by icezar-s         ###   ########.fr       */
+/*   Updated: 2025/12/06 19:41:04 by icezar-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@ int	ft_putcharc(char c)
 
 int	ft_putstrc(char *str)
 {
-  if (str == NULL)
-    return (ft_putstrc("(null)"));
+	if (str == NULL)
+		return (ft_putstrc("(null)"));
 	write(1, str, ft_strlen(str));
 	return ((int)ft_strlen(str));
 }
@@ -32,7 +32,7 @@ int	ft_putnbrc(int nbr)
 	int		len;
 
 	num = ft_itoa(nbr);
-	len = (int)ft_putstrc(num);
+	len = ft_putstrc(num);
 	free(num);
 	return (len);
 }
